pyproject.toml
[build-system]
build-backend = "pdm.backend"

requires = [ "pdm-backend" ]

[project]
name = "unusefulscripts"
version = "0.0.0"
requires-python = "==3.12.*"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "pillow>=10.4",
    "pypdf>=4.3.1",
    "pyside6>=6.7.2",
    "sdbus-notifications>=1.0.2",
]

urls.Homepage = "https://github.com/ZeeD/unusefulscripts"
urls.Repository = "https://github.com/ZeeD/unusefulscripts.git"

scripts.clocks = "clocks:main"
scripts.EstraiData = "estraidata:main"
scripts.prepend = "prepend:main"
scripts.SistemaNumeri = "sistemanumeri:main"
scripts.slacknews = "slacknews:main"
scripts.sub_offset_changer = "suboffsetchanger:main"
scripts.Televideo = "televideo:main"
scripts.TrovaRaggruppamenti = "trovaraggruppamenti:main"
scripts.unicodenames = "unicodenames:main"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "mypy>=1.10.1",
    "ruff>=0.5.0",
    "coverage>=7.5.4",
    "pyproject-fmt>=2.2.1",
    "types-beautifulsoup4>=4.12.0.20240511",
]

[tool.pdm.scripts]
_.env = { PYTHONPATH = "src:tests" }
lint = { composite = [
    "pyproject-fmt pyproject.toml",
    "ruff format {args:.}",
    "ruff check --exit-zero {args:.}",
    "dmypy run {args:.}",
] }
tests = { composite = [
    "coverage run -m unittest discover --verbose --locals --failfast --catch --start-directory {args:tests}",
    "coverage report",
    "coverage erase",
] }

[tool.ruff]
line-length = 80
src = [ "src", "tests" ]

fix = true
format.quote-style = "single"
format.skip-magic-trailing-comma = true
format.docstring-code-format = true
lint.select = [ "ALL" ]
lint.ignore = [
    "ANN1",
    # conflicts with formatter
    "COM812",
    # disabled by me
    "D1",
    # incompatibilities
    "D203",
    "D213",
    "ISC001",
    "PT009",
    "Q001",
    "Q002",
]
lint.flake8-annotations.allow-star-arg-any = true
lint.flake8-quotes.docstring-quotes = "single"
lint.flake8-quotes.inline-quotes = "single"
lint.flake8-quotes.multiline-quotes = "single"
lint.flake8-type-checking.quote-annotations = true
lint.isort.force-single-line = true

[tool.pyproject-fmt]
column_width = 80
indent = 4

[tool.coverage.run]
branch = true
source = [ "src", "tests" ]

[tool.coverage.report]
show_missing = true
sort = "Cover"
skip_empty = true

[tool.mypy]
mypy_path = [ "src", "tests" ]
strict = true
warn_unused_configs = true
